refactor: extract prompt functions to feedPrompt.ts and simplify interface
Extract buildSystemPrompt, buildHighlightContext, and resolveLanguage
into a dedicated feedPrompt.ts module, reducing generateFeed.ts from
442 to 354 lines. Remove unused language parameter from
CardGenerationService interface - language is already embedded in the
system prompt via buildLanguageInstruction.
