+++ packages/backend/convex/providers/cardGeneration.ts
@@ -25,7 +25,6 @@ export class AiSdkCardGenerator implements CardGenerationService {
     systemPrompt: string;
     userPrompt: string;
     cardCount: number;
-    language?: string;
   }): Promise<{ cards: Record<string, unknown>[]; usage: TokenUsage }> {
     const { output, usage } = await generateText({
       model: getAI().languageModel("premium"),
