+++ packages/backend/convex/providers/types.ts
@@ -5,12 +5,7 @@ export type TokenUsage = {
 };
 
 export interface CardGenerationService {
-  generateCards(opts: {
-    systemPrompt: string;
-    userPrompt: string;
-    cardCount: number;
-    language?: string;
-  }): Promise<{
+  generateCards(opts: { systemPrompt: string; userPrompt: string; cardCount: number }): Promise<{
     cards: Record<string, unknown>[];
     usage: TokenUsage;
   }>;
